+++ eTradeREST/src/main/java/edu/mum/service/CashAccountService.java
@@ -8,7 +8,7 @@ public interface CashAccountService {
 	
 	public void save(CashAccount cashAccount);
 	public List<CashAccount> findAll();
-	public CashAccount findOne(Long id);
+	public CashAccount findOne(Integer id);
 	public CashAccount update(CashAccount cashAccount);
 
 }
