+++ eTradeREST/src/main/java/edu/mum/service/ClientService.java
@@ -11,7 +11,7 @@ import edu.mum.domain.Client;
 public interface ClientService {
 	public void save(Client client);
 	public List<Client> findAll();
-	public Client findOne(Long id);
+	public Client findOne(Integer id);
 	public Client findByEmail(String email);
 	public Client update(Client client);
 	public void saveFull( Client client);
