+++ eTradeREST/src/main/java/edu/mum/service/UserService.java
@@ -8,7 +8,7 @@ public interface UserService {
 
 	public void save(User user);
 	public List<User> findAll();
-	public User findOne(Long id);
+	public User findOne(Integer id);
 	public User findByEmail(String email);
 	public User update(User user);
 	public void saveFull( User user);
